[Feature]: split admin routes
