+++ app/src/routes/_adminLayout/admin/(store)/coupons.analytics.lazy.tsx
@@ -1,7 +1,7 @@
 import UnderConstruction from "@/components/under-construction";
-import { createFileRoute } from "@tanstack/react-router";
+import { createLazyFileRoute } from "@tanstack/react-router";
 
-export const Route = createFileRoute("/_adminLayout/admin/(store)/coupons/analytics")({
+export const Route = createLazyFileRoute("/_adminLayout/admin/(store)/coupons/analytics")({
     component: RouteComponent,
 });
